Fixed compilation on cygwin 1.7 (Closes: #708).
git-svn-id: http://svn.aircrack-ng.org/trunk@1640 28c6078b-6c39-48e3-add9-af49d547ecab

+++ src/airbase-ng.c
@@ -1128,7 +1128,7 @@ int addESSIDfile(char* filename)
     {
         // trim trailing whitespace
         x = strlen(essid) - 1;
-        while (x >= 0 && isspace(essid[x]))
+        while (x >= 0 && isspace((int)essid[x]))
             essid[x--] = 0;
 
         if(strlen(essid))
